+++ src/main/java/roomescape/GeneralExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -9,6 +9,7 @@ public class GeneralExceptionHandler {
 
     @ExceptionHandler(IllegalArgumentException.class)
     public ResponseEntity<String> handleException(IllegalArgumentException e) {
+        System.out.println("[ERROR]" + e.getMessage());
         return ResponseEntity.badRequest().body("[ERROR]" + e.getMessage());
     }
 }
