+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,5 +1,6 @@
 package roomescape.controller;
 
+import jakarta.validation.Valid;
 import java.net.URI;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -29,7 +30,7 @@ public class ReservationController {
     }
 
     @PostMapping("/reservations")
-    public ResponseEntity<Reservation> addReservation(@RequestBody ReservationRequestDto request) {
+    public ResponseEntity<Reservation> addReservation(@RequestBody @Valid ReservationRequestDto request) {
         Reservation newReservation = reservationService.addReservation(request);
 
         return ResponseEntity.created(URI.create("/reservations/" + newReservation.getId())).body(newReservation);
